+++ pkg/inject/config.go
@@ -7,8 +7,8 @@ import (
 type ContainerVolumeMounts map[string][]corev1.VolumeMount
 
 type PatchConfig struct {
-	InitContainers        []corev1.Container      `yaml:"initContainers"`
-	Containers            []corev1.Container      `yaml:"containers"`
-	Volumes               []corev1.Volume         `yaml:"volumes"`
+	InitContainers        []corev1.Container    `yaml:"initContainers"`
+	Containers            []corev1.Container    `yaml:"containers"`
+	Volumes               []corev1.Volume       `yaml:"volumes"`
 	ContainerVolumeMounts ContainerVolumeMounts `yaml:"volumeMounts"`
 }