+++ pkg/inject/utils_test.go
@@ -4,36 +4,10 @@ import (
 	"bytes"
 	"encoding/json"
 	"io/ioutil"
-	"testing"
 	"text/template"
 
 	jsonpatch "github.com/evanphx/json-patch"
-	"github.com/stretchr/testify/assert"
 )
-
-func TestAuthenticatorSidecarInjection(t *testing.T) {
-	// Create the Admission Request (wrapped in an Admission Review) from the annotated
-	// pod fixture. The goal is to use the annotations as a signal to the sidecar-injector
-	// to mutate the Pod template spec.
-	req, err := newTestAdmissionRequest("./fixtures/authenticator-annotated-pod.json")
-	if !assert.NoError(t, err) {
-		return
-	}
-	// Read the expected Pod template spec fixture
-	expectedMod, err := ioutil.ReadFile("./fixtures/authenticator-mutated-pod.json")
-	if !assert.NoError(t, err) {
-		return
-	}
-
-	// Get the modified Pod template spec from the input .
-	mod, err := applyPatchToAdmissionRequest(req)
-	if !assert.NoError(t, err) {
-		return
-	}
-
-	// Assert that the modified Pod template spec should equal the expected Pod template spec
-	assert.JSONEq(t, string(expectedMod), string(mod))
-}
 
 // applyPatchToAdmissionRequest runs an AdmissionRequest (wrapped in an AdmissionReview)
 // through the sidecar-injector logic to extract a mutation patch, it then applies this
@@ -43,7 +17,13 @@ func applyPatchToAdmissionRequest(reviewRequestBytes []byte) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-	admissionRes := HandleAdmissionRequest(req)
+	admissionRes := HandleAdmissionRequest(
+		SidecarInjectorConfig{
+			SecretlessContainerImage:    "secretless-image",
+			AuthenticatorContainerImage: "authenticator-image",
+		},
+		req,
+	)
 
 	patch, err := jsonpatch.DecodePatch(admissionRes.Patch)
 	if err != nil {
@@ -59,7 +39,9 @@ func applyPatchToAdmissionRequest(reviewRequestBytes []byte) ([]byte, error) {
 // Admission Requests.
 func newTestAdmissionRequest(podTemplateSpecPath string) ([]byte, error) {
 	t := template.Must(
-		template.ParseFiles("./fixtures/authenticator-admission-request.tmpl.json"),
+		template.ParseFiles(
+			"./testdata/authenticator-admission-request.tmpl.json",
+		),
 	)
 
 	pod, err := ioutil.ReadFile(podTemplateSpecPath)